requirements.txt
streamlit
Pillow
google-generativeai
gtts
PyPDF2